fixed db models

+++ Models/Message.cs
@@ -4,7 +4,7 @@ namespace KeepThingsAPI.Models{
         public int id {get;set;}
         public int chat_id { get; set; }      
         public string message {get;set;}
-        public long sender_id { get; set; }
+        public int sender_id { get; set; }
         public long sent_timestamp {get;set;}
     }
 }
